requirement.txt
asgiref==3.2.10
attrs==20.2.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
distro==1.5.0
Django==3.1
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
idna==2.10
jsonschema==3.2.0
paramiko==2.7.2
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
texttable==1.6.3
urllib3==1.25.10
websocket-client==0.57.0
